containers/machine_statistics/requirements.txt
psycopg2-binary==2.9.10
Flask==3.1.0